ISSD-3152 Adding the BaseRestController
